Bugfixes, styling for favorites - prep for form util refactor
