+++ client/src/components/notepads/favorite_button.tsx
@@ -1,13 +1,12 @@
-import React, {useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import {connect} from "react-redux";
 import {createFavorite, deleteFavorite} from "../../actions/favorite_actions";
 import {RootState} from "../../reducers/root_reducer";
 import {Tooltip} from "antd";
 import {HeartOutlined, HeartFilled} from "@ant-design/icons";
-import {Notepad} from "../../interfaces";
 
 interface Props {
-    favorite: Notepad;
+    isFavorite: boolean;
     userId: number | null;
     notepadId: number;
     createFavorite: typeof createFavorite;
@@ -15,17 +14,18 @@ interface Props {
 }
 
 const FavoriteButton : React.FC<Props> = ({
-favorite,
+isFavorite,
 userId,
 notepadId,
 createFavorite,
 deleteFavorite}) => {
-    const [favoriteState, setFavoriteState] = useState<boolean>(!!favorite);
-    const [hoverState, setHoverState] = useState<boolean>(favoriteState);
+    const [favoriteState, setFavoriteState] = useState<boolean>(false);
+
+    useEffect(() => setFavoriteState(isFavorite), [isFavorite]);
 
     const handleClick = () => {
         if (!userId) return null;
-        if (!!favorite) {
+        if (isFavorite) {
             deleteFavorite({ UserID: userId, NotepadID: notepadId });
             setFavoriteState(false);
         } else {
@@ -34,15 +34,15 @@ deleteFavorite}) => {
         }
     }
 
-    const Icon = hoverState ? HeartFilled : HeartOutlined;
+    const Icon = favoriteState ? HeartFilled : HeartOutlined;
 
     return (
         <Tooltip title={userId ? 'Click to favorite!' : 'Log in to favorite!'}>
             <Icon onClick={handleClick}
-                  onMouseOverCapture={() => setHoverState(true)}
+                  onMouseOverCapture={() => setFavoriteState(true)}
                   onMouseLeave={() => {
-                      if (!favoriteState) {
-                          setHoverState(false);
+                      if (!isFavorite) {
+                          setFavoriteState(false);
                       }
                   }} />
         </Tooltip>
@@ -50,7 +50,7 @@ deleteFavorite}) => {
 };
 
 const mapStateToProps = (state : RootState, ownProps : { notepadId: number}) => ({
-    favorite: state.entities.favorites[ownProps.notepadId],
+    isFavorite: !!state.entities.favorites[ownProps.notepadId],
     userId: state.session.currentUserId
 });
